Move config implementations to a .cpp file.
