+++ UnleashedRecomp/app.cpp
@@ -7,6 +7,7 @@
 #include <patches/inspire_patches.h>
 #include <ui/game_window.h>
 #include <user/config.h>
+#include <user/paths.h>
 
 void App::Restart(std::vector<std::string> restartArgs)
 {
